Update upload file

+++ app/Classes/Upload.php
@@ -9,45 +9,65 @@ use Illuminate\Support\Str;
 
 class Upload
 {
-
-
     public static function uploadImage($image)
     {
-
         $imageSrc = env('IMAGE_SOURCE_PATH');
-        if ($image != null) {
-            $path = $imageSrc . date('Y') . '/' . date('m');
-            $originalName = pathinfo($image->getClientOriginalName(), PATHINFO_FILENAME);
-            $filename = $originalName . '_' . uniqid() . '.webp'; // Change the extension to .webp
+        $fileList = ['jpg', 'jpeg', 'png', 'webp', 'svg', 'gif', 'tiff', 'heic', 'raw'];
+        try {
+            $extension = strtolower($image->getClientOriginalExtension());
+            $originalName = pathinfo($image->getClientOriginalName(), PATHINFO_FILENAME) . '.' . $extension;
+            if ($image != null && in_array($extension, $fileList)) {
 
-            // Create the directory if it doesn't exist
-            if (!Storage::exists($path)) {
-                Storage::makeDirectory($path);
+                // Kiểm tra kích thước của ảnh
+                if ($image->getSize() > 5000000) {
+                    return [
+                        'status' => 'error',
+                        'message' => 'Dung lượng tệp ---> ' . $originalName . ' <--- không được vượt quá 5MB.'
+                    ];
+                }
+
+                $uuid = uniqid();
+                $path = $imageSrc . date('Y') . '/' . date('m');
+                $filename = Str::slug($originalName) . '_' . $uuid . '.webp'; // Change the extension to .webp
+
+                // Create the directory if it doesn't exist
+                if (!Storage::exists($path)) {
+                    Storage::makeDirectory($path);
+                }
+
+                // Optimize and convert the image to .webp
+                $img = Image::make($image)
+                    ->encode('webp', 90); // Encode to .webp with 80% quality
+
+                // Save the optimized image temporarily
+                $temporaryPath = storage_path('app/temp/' . $filename);
+                $img->save($temporaryPath);
+
+                // Optimize the image further using spatie/image-optimizer
+                $optimizerChain = OptimizerChainFactory::create();
+                $optimizerChain->optimize($temporaryPath);
+
+                // Move the optimized image to the final destination
+                $storedPath = $path . '/' . $filename;
+                Storage::put($storedPath, file_get_contents($temporaryPath));
+
+                // Remove temporary file
+                unlink($temporaryPath);
+                return [
+                    'status' => 'success',
+                    'message' => 'Tải lên ảnh thành công.'
+                ];
+            } else {
+                return [
+                    'status' => 'error',
+                    'message' => 'Định dạng tệp ---> ' . $originalName . ' <--- không hợp lệ. Chỉ chấp nhận các định dạng: JPG, JPEG, PNG, WEBP, SVG, GIF, TIFF, HEIC, RAW.'
+                ];
             }
-
-            // Optimize and convert the image to .webp
-            $img = Image::make($image)
-                ->encode('webp', 90); // Encode to .webp with 80% quality
-
-            // Save the optimized image temporarily
-            $temporaryPath = storage_path('app/temp/' . $filename);
-            $img->save($temporaryPath);
-
-            // Optimize the image further using spatie/image-optimizer
-            $optimizerChain = OptimizerChainFactory::create();
-            $optimizerChain->optimize($temporaryPath);
-
-            // Move the optimized image to the final destination
-            $storedPath = $path . '/' . $filename;
-            Storage::put($storedPath, file_get_contents($temporaryPath));
-
-            // Remove temporary file
-            unlink($temporaryPath);
-
-            // Remove 'public/uploads/photos' in accordance with AppServiceProvider
-            $newPath = Str::replaceFirst($imageSrc, 'images/', $storedPath);
-            return asset($newPath);
+        } catch (\Exception $e) {
+            return [
+                'status' => 'error',
+                'message' => 'Có lỗi từ server vui lòng thử lại.'
+            ];
         }
-        return null;
     }
 }
